pyproject.toml
[project]
name = "sabm-vllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
    "vllm>=0.9.1",
]
